MINOR: Preserve the base offset of the original record batch in V2
The previous code did not handle this correctly if a batch was
compacted more than once.

Also add test case for duplicate check after log cleaning and
improve various comments.

Author: Jason Gustafson <jason@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #3145 from hachikuji/minor-improve-base-sequence-docs

(cherry picked from commit 37433638271718344498d695d5da08db12c24eed)

+++ clients/src/main/java/org/apache/kafka/common/record/DefaultRecord.java
@@ -60,7 +60,7 @@ import static org.apache.kafka.common.utils.Utils.wrapNullable;
  *  ----------------
  *
  * The offset and timestamp deltas compute the difference relative to the base offset and
- * base timestamp of the log entry that this record is contained in.
+ * base timestamp of the batch that this record is contained in.
  */
 public class DefaultRecord implements Record {
 
@@ -102,7 +102,7 @@ public class DefaultRecord implements Record {
     }
 
     @Override
-    public long sequence() {
+    public int sequence() {
         return sequence;
     }
